mongoose adapter
