+++ packages/adapter-mongodb/README.md
@@ -1,13 +1,13 @@
-# Supabase adapter for Lucia
+# Mongoose (MongoDB) adapter for Lucia
 
-Adapter documentation: https://lucia-sveltekit.vercel.app/adapters/supabase
+Adapter documentation: https://lucia-sveltekit.vercel.app/adapters/mongoose
 
 Lucia documentation: https://lucia-sveltekit.vercel.app
 
 ## Installation
 
 ```
-npm install @lucia-sveltekit/adapter-supabase
+npm install @lucia-sveltekit/adapter-mongoose
 ```
 
 ## Testing
